displayed rides info
