+++ src/booking-list/BookingList.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { CalendarItems } from 'src/booking-list/CalendarItems';
+import { CalendarItemsWrapper } from 'src/booking-list/CalendarItemsWrapper';
 
 export const BookingList = () => {
   return (
     <div className={'bookingPage'}>
       <div className={'bookingWrapper'}>
         <div className={'bookingHeader cellDecoration'}>Booking list</div>
-        <CalendarItems />
+        <CalendarItemsWrapper />
       </div>
     </div>
   );
