+++ src/booking-list/CalendarItems.tsx
@@ -1,32 +1,17 @@
 import React from 'react';
-import { datesManager } from 'src/booking-list/datesManager';
-import { bookingItemsManager } from 'src/booking-list/BookingItemsManager';
-import { useBookingRides } from 'src/booking-list/useBookingRides';
+import { IUICalendarItems } from 'src/booking-list/types';
 
-export const CalendarItems = () => {
-  const requestRides = useBookingRides();
-
-  const datesMS = datesManager.getDatesMS();
-
-  const startEndDay = datesManager.getStartAndEndOfDays(datesMS);
-
-  const datesWithRides = bookingItemsManager.makeOneArrayFromTwo(
-    startEndDay,
-    requestRides,
-  );
-  console.log('obj', Object.entries(datesWithRides));
-  console.log(
-    'uiObj',
-    bookingItemsManager.calendarItemMapToUI(Object.entries(datesWithRides)),
-  );
-
-  const items = Object.entries(datesWithRides).map((info, index) => {
+export const CalendarItems = ({ datesRidesForUI }: IUICalendarItems) => {
+  const items = datesRidesForUI.map((item, index) => {
     return (
       <div className={'calendar'} key={index}>
+        <li className={'cellDecoration'}>{item.date}</li>
         <li className={'cellDecoration'}>
-          {datesManager.getFormattingDate(+info[0])}
+          <ul>
+            <li>{item.info.username}</li>
+            <li>{item.info.description}</li>
+          </ul>
         </li>
-        <li className={'cellDecoration'}>{}</li>
       </div>
     );
   });
